requirements.txt
Flask
Flask-Cors
requests
python-dotenv
google-generativeai
Pillow
pytest